moved reducers into seperate files and imported them into index.js
