+++ src/redux/reducers/movies.reducer.jsx
@@ -1,0 +1,11 @@
+// Used to store movies returned from the server
+const movies = (state = [], action) => {
+    switch (action.type) {
+        case 'SET_MOVIES':
+            return action.payload;
+        default:
+            return state;
+    }
+}
+
+export default movies;
