+++ src/redux/sagas/sagas.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import './index.css';
+import App from './components/App/App.js';
+import { createStore, combineReducers, applyMiddleware } from 'redux';
+// Provider allows us to use redux within our react app
+import { Provider } from 'react-redux';
+import logger from 'redux-logger';
+// Import saga middleware
+import createSagaMiddleware from 'redux-saga';
+import { takeEvery, put } from 'redux-saga/effects';
+import axios from 'axios';
